+++ CHANGELOG.md
@@ -17,6 +17,8 @@
     (prevents form double submit for example)
 - Added `:entity-path` key to sync options. When specified, will look up entity
   at path and use it for `:route-params` and `:params` of a request.
+- Made "reitit could not generate path" warning less noisy
+
 
 ## [0.13.2] 2020-09-06